Update VRED-VR-createNotes.py
Hi Simon,
Fixed the menu interchangeability issue: When the new menu button is pressed the already enabled button will be disabled automatically. Changed the name of the tool and added a new condition in the enable function to check the names of te tools starts with tools_. Additionally changed the name of the 'Multibuttonpad'.